Moved glue classes
